finish with gitignores

+++ scientific-computing/1list/5ex.jl
@@ -1,6 +1,6 @@
 # Stanisław Woźniak
 
-type = Float32
+type = Float64
 
 n = 5
